+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   * Sorting column names by table name on suggestions
   * Changed the icon on sugestions if it`s a view to make it visually different of a table
   * Fixed formatting with CRLF. Formatter issue [#3](https://github.com/mtxr/sqltools-formatter/issues/3)
+  * Allow users to set connection icons
+  * Add command `SQLTools.focusOnExplorer` to focus on explorer.
 
 * **Fixes**
   - Fixed history cutting some query parts on history explorer.